requirements/test.txt
-r prod.txt
django-debug-toolbar==5.0.1
flake8==7.1.1
flake8-quotes==3.4.0
pytest==8.3.4
pytest-django==4.9.0
